+++ test/form/samples/builtin-prototypes/object-expression/_config.js
@@ -1,3 +1,4 @@
 module.exports = {
-	description: 'Tree-shake known object prototype functions'
+	description: 'Tree-shake known object prototype functions',
+	expectedWarnings: ['EMPTY_BUNDLE']
 };