+++ test/form/samples/compact-multiple-imports/_config.js
@@ -5,6 +5,10 @@ module.exports = {
 			return id.startsWith('external');
 		},
 		output: {
+			globals: {
+				'external-3': 'external3',
+				'external-4': 'external4'
+			},
 			compact: true
 		}
 	}