+++ test/form/samples/deprecated/prune-pure-unused-import/_config.js
@@ -1,8 +1,9 @@
 module.exports = {
+	description: 'prunes pure unused external imports ([#1352])',
+	expectedWarnings: ['EMPTY_BUNDLE'],
 	options: {
 		strictDeprecations: false,
 		external: ['external', 'other'],
 		treeshake: { pureExternalModules: true }
-	},
-	description: 'prunes pure unused external imports ([#1352])'
+	}
 };