+++ test/form/samples/deprecated/transform-bundle-plugin/_config.js
@@ -1,10 +1,11 @@
 module.exports = {
 	description: 'allows plugins to transform bundle',
+	expectedWarnings: ['DEPRECATED_FEATURE'],
 	options: {
 		strictDeprecations: false,
 		plugins: [
 			{
-				transformBundle(code) {
+				transformBundle() {
 					return '/* first plugin */';
 				}
 			},