+++ test/form/samples/guessed-global-names/_config.js
@@ -1,0 +1,15 @@
+const { resolve } = require('path');
+
+module.exports = {
+	description: 'guesses global names if necessary',
+	expectedWarnings: ['MISSING_GLOBAL_NAME'],
+	options: {
+		external: [
+			'unchanged',
+			'changed',
+			'special-character',
+			'with/slash',
+			resolve(__dirname, 'relative.js')
+		]
+	}
+};