+++ test/form/samples/namespace-tostringtag/_config.js
@@ -1,5 +1,6 @@
 module.exports = {
 	description: 'namespace toStringTag support',
+	expectedWarnings: ['CIRCULAR_DEPENDENCY'],
 	options: {
 		output: {
 			name: 'iife',