+++ test/form/samples/reexports-from-external/_config.js
@@ -2,6 +2,9 @@ module.exports = {
 	description: 're-exports * from external module (#791)',
 	options: {
 		external: ['external'],
-		output: { name: 'myBundle' }
+		output: {
+			globals: { external: 'external' },
+			name: 'myBundle'
+		}
 	}
 };