+++ test/form/samples/render-chunk-plugin-sourcemaps/_config.js
@@ -1,13 +1,14 @@
 module.exports = {
 	description:
 		'supports returning undefined source maps from render chunk hooks, when source maps are enabled',
+	expectedWarnings: ['SOURCEMAP_BROKEN'],
 	options: {
 		output: {
 			sourcemap: true
 		},
 		plugins: [
 			{
-				renderChunk(code) {
+				renderChunk() {
 					return '/* first plugin */';
 				}
 			},