+++ test/form/samples/self-calling-function/_config.js
@@ -1,3 +1,4 @@
 module.exports = {
-	description: 'discards a self-calling function without side-effects'
+	description: 'discards a self-calling function without side-effects',
+	expectedWarnings: ['EMPTY_BUNDLE']
 };