+++ test/form/samples/side-effect-r/_config.js
@@ -1,4 +1,5 @@
 module.exports = {
 	description:
-		'discards unused function expression assigned to a variable that calls itself and a global'
+		'discards unused function expression assigned to a variable that calls itself and a global',
+	expectedWarnings: ['EMPTY_BUNDLE']
 };