+++ test/form/samples/side-effect-s/_config.js
@@ -1,4 +1,5 @@
 module.exports = {
 	description:
-		'discards unused function expression assigned to a variable that calls itself and has side effects'
+		'discards unused function expression assigned to a variable that calls itself and has side effects',
+	expectedWarnings: ['EMPTY_BUNDLE']
 };