+++ test/form/samples/side-effect-with-plusplus-expression/_config.js
@@ -1,3 +1,5 @@
 module.exports = {
-	description: 'correctly drop side-effect free statements in presence of update expression (#1564)'
+	description:
+		'correctly drop side-effect free statements in presence of update expression (#1564)',
+	expectedWarnings: ['EMPTY_BUNDLE']
 };