+++ test/form/samples/unused-import/_config.js
@@ -1,6 +1,9 @@
 module.exports = {
 	description: 'excludes unused imports ([#595])',
 	options: {
-		external: ['external']
+		external: ['external'],
+		output: {
+			globals: { external: 'external' }
+		}
 	}
 };