+++ test/function/index.js
@@ -52,7 +52,6 @@ runTestSuiteWithSamples('function', path.resolve(__dirname, 'samples'), (dir, co
 		() => {
 			if (config.show) console.group(path.basename(dir));
 			if (config.before) config.before();
-
 			process.chdir(dir);
 			const warnings = [];