+++ test/function/samples/warnings-to-string/main.js
@@ -1,0 +1,1 @@
+const removed = 1;